Fix: MDSplus.wsgi use treecontext and added config scripts for easy apache integration; moves js and html
