+++ deploy/packaging/redhat/python.noarch
@@ -7,16 +7,11 @@
 ./usr/local/mdsplus/mdsobjects/python/connection.py
 ./usr/local/mdsplus/mdsobjects/python/descriptor.py
 ./usr/local/mdsplus/mdsobjects/python/event.py
-./usr/local/mdsplus/mdsobjects/python/html
-./usr/local/mdsplus/mdsobjects/python/html/scope.html
-./usr/local/mdsplus/mdsobjects/python/js
-./usr/local/mdsplus/mdsobjects/python/js/mdsScope.js
 ./usr/local/mdsplus/mdsobjects/python/magic.py
 ./usr/local/mdsplus/mdsobjects/python/mdsExceptions.py
 ./usr/local/mdsplus/mdsobjects/python/mdsarray.py
 ./usr/local/mdsplus/mdsobjects/python/mdsdata.py
 ./usr/local/mdsplus/mdsobjects/python/mdsdcl.py
-./usr/local/mdsplus/mdsobjects/python/mdsplus_wsgi.py
 ./usr/local/mdsplus/mdsobjects/python/mdsscalar.py
 ./usr/local/mdsplus/mdsobjects/python/modpython.py
 ./usr/local/mdsplus/mdsobjects/python/scope.py
@@ -73,9 +68,13 @@
 ./usr/local/mdsplus/mdsobjects/python/widgets/mdsplusxdboxbutton.py
 ./usr/local/mdsplus/mdsobjects/python/wsgi
 ./usr/local/mdsplus/mdsobjects/python/wsgi/__init__.py
+./usr/local/mdsplus/mdsobjects/python/wsgi/conf
+./usr/local/mdsplus/mdsobjects/python/wsgi/conf/mdsplus.conf
+./usr/local/mdsplus/mdsobjects/python/wsgi/conf/mdsplus.wsgi
 ./usr/local/mdsplus/mdsobjects/python/wsgi/colors1.tbl
 ./usr/local/mdsplus/mdsobjects/python/wsgi/do1darray.py
 ./usr/local/mdsplus/mdsobjects/python/wsgi/do1dsignal.py
+./usr/local/mdsplus/mdsobjects/python/wsgi/doDir.py
 ./usr/local/mdsplus/mdsobjects/python/wsgi/doEvent.py
 ./usr/local/mdsplus/mdsobjects/python/wsgi/doFrame.py
 ./usr/local/mdsplus/mdsobjects/python/wsgi/doGetnid.py
@@ -85,3 +84,7 @@
 ./usr/local/mdsplus/mdsobjects/python/wsgi/doPlot.py
 ./usr/local/mdsplus/mdsobjects/python/wsgi/doScope.py
 ./usr/local/mdsplus/mdsobjects/python/wsgi/doTreepath.py
+./usr/local/mdsplus/mdsobjects/python/wsgi/html
+./usr/local/mdsplus/mdsobjects/python/wsgi/html/scope.html
+./usr/local/mdsplus/mdsobjects/python/wsgi/js
+./usr/local/mdsplus/mdsobjects/python/wsgi/js/mdsScope.js
