+++ mdsobjects/python/wsgi/conf/mdsplus.conf
@@ -1,0 +1,26 @@
+###************************************************************
+### MDSplus.wsgi config script (apache2)                      *
+###                                                           *
+### adapt server info and place this file in config folder    *
+### e.g. '/etc/apache2/mods-enabled/'                          *
+###************************************************************
+
+<VirtualHost *:80>
+
+    ServerName  www.mdsplus.local.com
+    ServerAlias mdsplus.local.com
+    ServerAdmin admin@local.com
+
+    WSGIScriptAlias /mdsplusWsgi /var/www/mdsplusWsgi/mdsplus.wsgi
+
+    <Directory "/var/www/mdsplusWsgi">
+        <IfVersion < 2.4>
+            Order allow,deny
+            Allow from all
+        </IfVersion>
+        <IfVersion >= 2.4>
+            Require all granted
+        </IfVersion>
+    </Directory>
+
+</VirtualHost>
