+++ mdsobjects/python/wsgi/doEvent.py
@@ -31,12 +31,11 @@ def noCache(response_headers):
     response_headers.append(('expires','Sat, 26 Jul 1997 09:00:00 GMT'))
     response_headers.append(('Pragma','no-cache'))
 
+class myevent(Event):
+    def run(self):
+        self.cancel()
+
 def doEvent(self):
-
-    class myevent(Event):
-        def run(self):
-            self.cancel()
-
     status = '200 OK'
     response_headers=list()
     noCache(response_headers)
