+++ mdsobjects/python/wsgi/doFrame.py
@@ -23,7 +23,7 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 
-from MDSplus import makeData,Data,Tree,Float64Array
+from MDSplus import DATA,Float64Array,tdi
 import sys
 
 def doFrame(self):
@@ -43,12 +43,14 @@ def doFrame(self):
     response_headers.append(('Content-Type','application/octet-stream'))
 
     t = None
+    _tdi = tdi
     if 'tree' in self.args:
-        Tree.usePrivateCtx()
         try:
-            t = Tree(self.args['tree'][-1],int(self.args['shot'][-1].split(',')[0]))
+            t = self.openTree(self.args['tree'][-1],self.args['shot'][-1].split(',')[0])
         except:
             response_headers.append(('ERROR','Error opening tree'))
+        else:
+            _tdi = t.tdiExecute
 
     for name in ('title','xlabel','ylabel'):
         getStringExp(self,name,response_headers,t)
@@ -60,43 +62,11 @@ def doFrame(self):
 
     expr = self.args['y'][-1]
     try:
-        if t != None:
-            sig = t.tdiExecute('GetSegment(' + expr + ',' + frame_idx + ')')
-        else:
-            sig = Data.execute('GetSegment(' + expr + ',' + frame_idx + ')')
-        frame_data = makeData(sig.data())
+        sig = _tdi('GetSegment(' + expr + ',' + frame_idx + ')')
+        frame_data = DATA(sig).evaluate()
     except Exception:
         response_headers.append(('ERROR','Error evaluating expression: "%s", error: %s' % (expr,sys.exc_info())))
 
-    if 'init' in self.args:
-        if 'x' in self.args:
-            expr = self.args['x'][-1]
-            try:
-                if t == None:
-                    times = Data.execute(expr)
-                else:
-                    times = t.tdiExecute(expr)
-                times = makeData(times.data())
-            except Exception:
-                response_headers.append(('ERROR','Error evaluating expression: "%s", error: %s' % (expr,sys.exc_info())))
-        else:
-            try:
-                #times = Data.execute('dim_of(' + expr + ')')
-                times = list()
-                if t == None:
-                    numSegments = Data.execute('GetNumSegments(' + expr + ')').data()
-                else:
-                   numSegments = t.tdiExecute('GetNumSegments(' + expr + ')').data()
-                for i in range(0, numSegments):
-                    if t == None:
-                        times.append(Data.execute('GetSegmentLimits(' + expr + ',' + str(i) + ')').data()[0])
-                    else:
-                        times.append(t.tdiExecute('GetSegmentLimits(' + expr + ',' + str(i) + ')').data()[0])
-                
-                times = Float64Array(times)
-            except Exception:
-                response_headers.append(('ERROR','Error getting x axis of: "%s", error: %s' % (expr,sys.exc_info())))
-    
     response_headers.append(('FRAME_WIDTH',str(sig.getShape()[0])))
     response_headers.append(('FRAME_HEIGHT',str(sig.getShape()[1])))
     response_headers.append(('FRAME_BYTES_PER_PIXEL',str(frame_data.data().itemsize)))
@@ -105,6 +75,23 @@ def doFrame(self):
     output = str(frame_data.data().data)
 
     if 'init' in self.args:
+        if 'x' in self.args:
+            expr = self.args['x'][-1]
+            try:
+                times = DATA(_tdi(expr)).evaulate()
+            except Exception:
+                response_headers.append(('ERROR','Error evaluating expression: "%s", error: %s' % (expr,sys.exc_info())))
+        else:
+            try:
+                #times = Data.execute('dim_of(' + expr + ')')
+                times = list()
+                numSegments = _tdi('GetNumSegments(' + expr + ')').data()
+                for i in range(0, numSegments):
+                    times.append(_tdi('GetSegmentLimits(' + expr + ',' + str(i) + ')').data()[0])
+                times = Float64Array(times)
+            except Exception:
+                response_headers.append(('ERROR','Error getting x axis of: "%s", error: %s' % (expr,sys.exc_info())))
+
         response_headers.append(('TIMES_DATATYPE',times.__class__.__name__))
         response_headers.append(('TIMES_LENGTH',str(len(times))))
         output = output + str(times.data().data)
